envelope: move files
Move the Envelope code into libenvelope.
Standardise the naming of the files/objects: wdata.c, window.c
