+++ compose/functions.c
@@ -48,6 +48,7 @@
 #include "lib.h"
 #include "attach/lib.h"
 #include "browser/lib.h"
+#include "envelope/lib.h"
 #include "index/lib.h"
 #include "menu/lib.h"
 #include "ncrypt/lib.h"
@@ -56,7 +57,6 @@
 #include "attach_data.h"
 #include "commands.h"
 #include "context.h"
-#include "env_data.h"
 #include "hook.h"
 #include "mutt_header.h"
 #include "mutt_logging.h"
