+++ compose/lib.h
@@ -33,8 +33,6 @@
  * | compose/cbar_data.c    | @subpage compose_cbar_data     |
  * | compose/compose.c      | @subpage compose_dialog        |
  * | compose/config.c       | @subpage compose_config        |
- * | compose/env_data.c     | @subpage compose_env_data      |
- * | compose/envelope.c     | @subpage compose_envelope      |
  * | compose/functions.c    | @subpage compose_functions     |
  * | compose/shared_data.c  | @subpage compose_shared_data   |
  */
