+++ envelope/window.c
@@ -1,6 +1,6 @@
 /**
  * @file
- * Compose Envelope
+ * Envelope Window
  *
  * @authors
  * Copyright (C) 2021 Richard Russon <rich@flatcap.org>
@@ -21,15 +21,15 @@
  */
 
 /**
- * @page compose_envelope Envelope Window
+ * @page envelope_window Envelope Window
  *
- * The Compose Envelope Window displays the header fields of an email.
+ * The Envelope Window displays the header fields of an email.
  *
  * ## Windows
  *
- * | Name                    | Type      | See Also          |
- * | :---------------------- | :-------- | :---------------- |
- * | Compose Envelope Window | WT_CUSTOM | compose_env_new() |
+ * | Name            | Type      | See Also          |
+ * | :-------------- | :-------- | :---------------- |
+ * | Envelope Window | WT_CUSTOM | env_window_new()  |
  *
  * **Parent**
  * - @ref compose_dialog
@@ -39,10 +39,9 @@
  * None.
  *
  * ## Data
- * - #ComposeEnvelopeData
+ * - #EnvelopeWindowData
  *
- * The Compose Envelope Window stores its data (#ComposeEnvelopeData) in
- * MuttWindow::wdata.
+ * The Envelope Window stores its data (#EnvelopeWindowData) in MuttWindow::wdata.
  *
  * ## Events
  *
@@ -71,9 +70,10 @@
 #include "gui/lib.h"
 #include "color/lib.h"
 #include "ncrypt/lib.h"
-#include "env_data.h"
+#include "../compose/private.h"
+#include "../compose/shared_data.h"
 #include "options.h"
-#include "shared_data.h"
+#include "wdata.h"
 #ifdef ENABLE_NLS
 #include <libintl.h>
 #endif
@@ -83,6 +83,51 @@
 
 /// Maximum number of rows to use for the Headers: field
 #define MAX_USER_HDR_ROWS 5
+
+int HeaderPadding[HDR_ATTACH_TITLE] = { 0 };
+int MaxHeaderWidth = 0;
+
+const char *const Prompts[] = {
+  /* L10N: Compose menu field.  May not want to translate. */
+  N_("From: "),
+  /* L10N: Compose menu field.  May not want to translate. */
+  N_("To: "),
+  /* L10N: Compose menu field.  May not want to translate. */
+  N_("Cc: "),
+  /* L10N: Compose menu field.  May not want to translate. */
+  N_("Bcc: "),
+  /* L10N: Compose menu field.  May not want to translate. */
+  N_("Subject: "),
+  /* L10N: Compose menu field.  May not want to translate. */
+  N_("Reply-To: "),
+  /* L10N: Compose menu field.  May not want to translate. */
+  N_("Fcc: "),
+#ifdef MIXMASTER
+  /* L10N: "Mix" refers to the MixMaster chain for anonymous email */
+  N_("Mix: "),
+#endif
+  /* L10N: Compose menu field.  Holds "Encrypt", "Sign" related information */
+  N_("Security: "),
+  /* L10N: This string is used by the compose menu.
+     Since it is hidden by default, it does not increase the indentation of
+     other compose menu fields.  However, if possible, it should not be longer
+     than the other compose menu fields.  Since it shares the row with "Encrypt
+     with:", it should not be longer than 15-20 character cells.  */
+  N_("Sign as: "),
+#ifdef USE_AUTOCRYPT
+  // L10N: The compose menu autocrypt line
+  N_("Autocrypt: "),
+#endif
+#ifdef USE_NNTP
+  /* L10N: Compose menu field.  May not want to translate. */
+  N_("Newsgroups: "),
+  /* L10N: Compose menu field.  May not want to translate. */
+  N_("Followup-To: "),
+  /* L10N: Compose menu field.  May not want to translate. */
+  N_("X-Comment-To: "),
+#endif
+  N_("Headers: "),
+};
 
 #ifdef USE_AUTOCRYPT
 static const char *const AutocryptRecUiFlags[] = {
@@ -108,6 +153,59 @@ static const char *const AutocryptRecUiFlags[] = {
   N_("Yes"),
 };
 #endif
+
+/**
+ * calc_header_width_padding - Calculate the width needed for the compose labels
+ * @param idx      Store the result at this index of HeaderPadding
+ * @param header   Header string
+ * @param calc_max If true, calculate the maximum width
+ */
+static void calc_header_width_padding(int idx, const char *header, bool calc_max)
+{
+  int width;
+
+  HeaderPadding[idx] = mutt_str_len(header);
+  width = mutt_strwidth(header);
+  if (calc_max && (MaxHeaderWidth < width))
+    MaxHeaderWidth = width;
+  HeaderPadding[idx] -= width;
+}
+
+/**
+ * init_header_padding - Calculate how much padding the compose table will need
+ *
+ * The padding needed for each header is strlen() + max_width - strwidth().
+ *
+ * calc_header_width_padding sets each entry in HeaderPadding to strlen -
+ * width.  Then, afterwards, we go through and add max_width to each entry.
+ */
+static void init_header_padding(void)
+{
+  static bool done = false;
+
+  if (done)
+    return;
+  done = true;
+
+  for (int i = 0; i < HDR_ATTACH_TITLE; i++)
+  {
+    if (i == HDR_CRYPTINFO)
+      continue;
+    calc_header_width_padding(i, _(Prompts[i]), true);
+  }
+
+  /* Don't include "Sign as: " in the MaxHeaderWidth calculation.  It
+   * doesn't show up by default, and so can make the indentation of
+   * the other fields look funny. */
+  calc_header_width_padding(HDR_CRYPTINFO, _(Prompts[HDR_CRYPTINFO]), false);
+
+  for (int i = 0; i < HDR_ATTACH_TITLE; i++)
+  {
+    HeaderPadding[i] += MaxHeaderWidth;
+    if (HeaderPadding[i] < 0)
+      HeaderPadding[i] = 0;
+  }
+}
 
 /**
  * calc_address - Calculate how many rows an AddressList will need
@@ -875,6 +973,8 @@ static int env_window_observer(struct NotifyCallback *nc)
  */
 struct MuttWindow *compose_env_new(struct ComposeSharedData *shared, struct Buffer *fcc)
 {
+  init_header_padding();
+
   struct MuttWindow *win_env =
       mutt_window_new(WT_CUSTOM, MUTT_WIN_ORIENT_VERTICAL, MUTT_WIN_SIZE_FIXED,
                       MUTT_WIN_SIZE_UNLIMITED, HDR_ATTACH_TITLE - 1);
